applied #13933 - Simplify QuadBuckets and improve memory footprint (patch by Gerd Petermann, minor modifications)
git-svn-id: http://josm.openstreetmap.de/svn/trunk@11237 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/data/coor/QuadTiling.java
@@ -74,10 +74,10 @@ public final class QuadTiling {
      * @return quad tiling index for given coordinates and level.
      * @since 6171
      */
-    public static int index(final double lat, final double lon, final int level) {
+    public static byte index(final double lat, final double lon, final int level) {
         long x = lon2x(lon);
         long y = lat2y(lat);
         int shift = NR_LEVELS-level-1;
-        return (int) ((x >> shift & 1) * 2 + (y >> shift & 1));
+        return (byte) ((x >> shift & 1) * 2 + (y >> shift & 1));
     }
 }
